+++ Changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Refactored the method FirstAgendaService::getCommitteesInOrganizations and 
+  FirstAgendaService::getAllCommitteesAvailable to use private method.
+
 ## [v0.3.1]
 
 ### Changed