+++ Changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.3.2]
+
 ### Added
 - Added Presentation, Documents and ItemDecision to ApiAgendaItem